+++ packages/f-header/test/specs/component/f-header.component.spec.js
@@ -1,11 +1,12 @@
 import HeaderComponent from '../../../test-utils/component-objects/f-header.component';
+import forEach from 'mocha-each';
+const path = '/iframe.html?id=components-organisms--' // storybook url for all components - could move to config.
+const offers = '&knob-Show%20offers%20link=true'
+const delivery = '&knob-Show%20delivery%20enquiry=true'
 
 describe('f-header component tests', () => {
-    before(() => {
-        browser.url('?path=/story/components-organisms--header-component');
-        HeaderComponent.selectAddOns();
-        browser.switchToFrame(0);
-        HeaderComponent.waitForHeader();
+    beforeEach(() => {
+        browser.url(`${path}header-component${offers}${delivery}`);
     });
 
     it('should display the f-header component', () => {
@@ -13,41 +14,41 @@ describe('f-header component tests', () => {
         expect(HeaderComponent.isLogoDisplayed()).toBe(true);
     });
 
+    it('should only show the default navigation fields', () => {
+        // Act
+        browser.url(`${path}header-component`);
+
+        // Assert
+        expect(HeaderComponent.isFieldLinkDisplayed('offers')).toBe(false); 
+        expect(HeaderComponent.isFieldLinkDisplayed('help')).toBe(true);
+    })
+
+    forEach(['offers', 'help', 'delivery', 'userAccount'])
+    .it('should show all the links and icons', field => {
+        // Assert
+        expect(HeaderComponent.isFieldLinkDisplayed(field)).toBe(true);
+    });
+
     it('should alter visibility of navbar depending on window size', () => {
         // Act
         browser.setWindowSize(500, 500);
 
         // Assert
-        expect(HeaderComponent.isMobileNavigationVisible()).toBe(true);
-        expect(HeaderComponent.isHelpLinkDisplayed()).toBe(false);
-        expect(HeaderComponent.isLoginLinkDisplayed()).toBe(false);
-
-        // Act
-        browser.setWindowSize(1000, 1000);
-
-        // Assert
-        expect(HeaderComponent.isMobileNavigationVisible()).toBe(false);
-        expect(HeaderComponent.isHelpLinkDisplayed()).toBe(true);
-        expect(HeaderComponent.isLoginLinkDisplayed()).toBe(true);
+        expect(HeaderComponent.isMobileNavigationBarVisible()).toBe(true);
+        expect(HeaderComponent.isMobileOffersIconDisplayed()).toBe(true);
+        expect(HeaderComponent.isFieldLinkDisplayed('help')).toBe(false);
     });
 
-    it('should change url when help link is clicked', () => {
+    it('should change url when help-link is clicked', () => {
         // Act
+        browser.setWindowSize(1000, 1000);
         HeaderComponent.clickHelpLink();
 
         // Assert
         expect(browser.getUrl()).toContain("http://localhost:8080/help");
     });
 
-    it.skip('should change url when login link is clicked', () => {
-        // Act
-        HeaderComponent.clickLoginLink();
-
-        // Assert
-        expect(browser.getUrl()).toContain("/account/login");
-    });
-
-    it.skip('should change the url to offers when offers link is clicked', () => {
+    it('should change the url to offers when offers link is clicked', () => {
         // Act
         HeaderComponent.clickOffersLink();
 
@@ -55,32 +56,16 @@ describe('f-header component tests', () => {
         expect(browser.getUrl()).toContain("/offers");
     });
 
-    it.skip('should display the "For You" icon for mobile', () => {
+    forEach(['offers', 'help', 'delivery', 'userAccount'])
+    .it.skip('should show navigation fields when mobile-navigation has been opened', field => {
         // Act
-        browser.setWindowSize(500, 500);
-
-        // Assert
-        expect(HeaderComponent.isMobileOffersIconDisplayed()).toBe(true);
-    });
-    // Skip until we have solved mobile icon
-    it.skip('should display expected offers link depending on viewport-size', () => {
-
-        // Assert
-        expect(HeaderComponent.isNavigationOffersLinkDisplayed()).toBe(true);
-        expect(HeaderComponent.isMobileOffersIconDisplayed()).toBe(false);
-
-        // Act
-        // resize to mobile
-        browser.setWindowSize(500, 500);
-
-        expect(HeaderComponent.isNavigationOffersLinkDisplayed()).toBe(false);
-        expect(HeaderComponent.isMobileOffersIconDisplayed()).toBe(true);
-
-        // Act
+        browser.setWindowSize(500, 1000);
         HeaderComponent.openMobileNavigation();
 
         // Assert
-        expect(HeaderComponent.isNavigationOffersLinkDisplayed()).toBe(true);
-        expect(HeaderComponent.isMobileOffersIconDisplayed()).toBe(true);
+        expect(HeaderComponent.isFieldLinkDisplayed(field)).toBe(true);
+
+         //Skipping for now - delivery is missing when nav is changed from web to mobile. 
+         //Will ask a front-end dev to confirm.
     });
 });