+++ packages/f-header/CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.4.0
+------------------------------
+*December 18, 2020*
+
+### Changed
+- Updated Webdriver test-utils to include `selector` file.
+- Refactored component tests.
+
 
 v4.3.0
 ------------------------------